Added initial Windows support!
Currently only tested on XP and requires Firefox to be started as an
Administrator.

MozPopen build stuff is a bit of a mess so beware.

Added alternate backend to mozpopen that uses Boost.Process.
http://www.highscore.de/cpp/process/
Should work on all platforms, but just using on Windows for now due to
time constraints.